+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/demandandcapacitynotification/logic/service/DemandAndCapacityNotifcationRequestApiService.java
@@ -20,13 +20,11 @@
 
 package org.eclipse.tractusx.puris.backend.demandandcapacitynotification.logic.service;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.extern.slf4j.Slf4j;
-import java.util.Date;
-import java.util.UUID;
 import javax.management.openmbean.KeyAlreadyExistsException;
+
 import org.eclipse.tractusx.puris.backend.common.edc.logic.service.EdcAdapterService;
+import org.eclipse.tractusx.puris.backend.common.industrycore.IndustryCoreMessageContext;
+import org.eclipse.tractusx.puris.backend.common.industrycore.IndustryCoreMessageService;
 import org.eclipse.tractusx.puris.backend.demandandcapacitynotification.domain.model.OwnDemandAndCapacityNotification;
 import org.eclipse.tractusx.puris.backend.demandandcapacitynotification.domain.model.ReportedDemandAndCapacityNotification;
 import org.eclipse.tractusx.puris.backend.demandandcapacitynotification.logic.adapter.DemandAndCapacityNotificationSammMapper;
@@ -36,6 +34,10 @@ import org.eclipse.tractusx.puris.backend.masterdata.logic.service.PartnerServic
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.fasterxml.jackson.databind.JsonNode;
+
+import lombok.extern.slf4j.Slf4j;
+
 @Service
 @Slf4j
 public class DemandAndCapacityNotifcationRequestApiService {
@@ -44,14 +46,11 @@ public class DemandAndCapacityNotifcationRequestApiService {
     @Autowired
     private ReportedDemandAndCapacityNotificationService reportedDemandAndCapacityNotificationService;
     @Autowired
+    private IndustryCoreMessageService messageService;
+    @Autowired
     private EdcAdapterService edcAdapterService;
     @Autowired
     private DemandAndCapacityNotificationSammMapper sammMapper;
-    @Autowired
-    private ObjectMapper objectMapper;
-
-    public static final String DEMAND_AND_CAPACITY_NOTIFICATION_CONTEXT = "CX-DemandAndCapacityNotificationAPI-Receive:2.0.0";
-    public static final String MESSAGE_HEADER_VERSION = "3.0.0";
 
     public ReportedDemandAndCapacityNotification handleIncomingNotification(String bpnl, DemandAndCapacityNotificationSamm samm) {
         Partner partner = partnerService.findByBpnl(bpnl);
@@ -94,20 +93,7 @@ public class DemandAndCapacityNotifcationRequestApiService {
     }
 
     private JsonNode createNotificationRequestBody(OwnDemandAndCapacityNotification notification) {
-        var ownPartnerEntity = partnerService.getOwnPartnerEntity();
-        var body = objectMapper.createObjectNode();
-        var header = objectMapper.createObjectNode();
-        body.set("header", header);
-        header.put("senderBpn", ownPartnerEntity.getBpnl());
-        header.put("receiverBpn", notification.getPartner().getBpnl());
-        header.put("context", DEMAND_AND_CAPACITY_NOTIFICATION_CONTEXT);
-        header.put("messageId", UUID.randomUUID().toString());
-        header.put("sentDateTime", new Date().toString());
-        header.put("version", MESSAGE_HEADER_VERSION);
-        var content = objectMapper.createObjectNode();
-        body.set("content", content);
         var samm = sammMapper.ownNotificationToSamm(notification);
-        content.set("demandAndCapacityNotification", objectMapper.convertValue(samm, JsonNode.class));
-        return body;
+        return messageService.createMessage(notification.getPartner(), IndustryCoreMessageContext.DEMAND_AND_CAPACITY_NOTIFICATION_CONTEXT, samm);
     }
 }